+++ domain/src/main/java/com/androiddev/domain/repository/notification/NotificationRepository.kt
@@ -5,5 +5,5 @@ import com.androiddev.domain.util.Resource
 import kotlinx.coroutines.flow.Flow
 
 interface NotificationRepository {
-    suspend fun getNotifications(notificationId:Int?,notificationDate:String?): Flow<Resource<List<NotificationItem>>>
+    suspend fun getNotifications(notificationId:Long?,notificationDate:String?): Flow<Resource<List<NotificationItem>>>
 }
